chore: refactoring ExploreSyncWorker and database for easier testing
