+++ features/exploredash/build.gradle
@@ -54,6 +54,7 @@ dependencies {
     implementation "androidx.core:core-ktx:$jetpackVersion"
     implementation "androidx.appcompat:appcompat:$appCompatVersion"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-play-services:$coroutines_version"
+    implementation "androidx.work:work-runtime-ktx:$workRuntimeVersion"
 
     // Architecture
     implementation "androidx.navigation:navigation-fragment-ktx:$navigationVersion"
@@ -64,6 +65,7 @@ dependencies {
     implementation "com.google.dagger:hilt-android:$hiltVersion"
     kapt "com.google.dagger:hilt-compiler:$hiltVersion"
     kapt 'androidx.hilt:hilt-compiler:1.0.0'
+    implementation "androidx.hilt:hilt-work:$hiltWorkVersion"
 
     // UI
     implementation "com.google.android.material:material:$materialVersion"
