+++ features/exploredash/src/main/java/org/dash/wallet/features/exploredash/ExploreSyncWorker.kt
@@ -1,0 +1,111 @@
+/*
+ * Copyright 2022 Dash Core Group.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package org.dash.wallet.features.exploredash
+
+import android.annotation.SuppressLint
+import android.content.Context
+import androidx.hilt.work.HiltWorker
+import androidx.work.CoroutineWorker
+import androidx.work.WorkerParameters
+import com.google.firebase.FirebaseNetworkException
+import dagger.assisted.Assisted
+import dagger.assisted.AssistedInject
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import org.dash.wallet.common.Configuration
+import org.dash.wallet.common.services.analytics.AnalyticsService
+import org.dash.wallet.features.exploredash.repository.ExploreRepository
+import org.dash.wallet.features.exploredash.repository.DataSyncStatusService
+import org.slf4j.LoggerFactory
+import java.util.*
+import kotlin.system.measureTimeMillis
+
+@HiltWorker
+class ExploreSyncWorker @AssistedInject constructor(
+    @Assisted private val appContext: Context,
+    @Assisted workerParams: WorkerParameters,
+    private val analytics: AnalyticsService,
+    private val exploreRepository: ExploreRepository,
+    private val syncStatus: DataSyncStatusService,
+    private val config: Configuration
+): CoroutineWorker(appContext, workerParams) {
+    companion object {
+        private val log = LoggerFactory.getLogger(ExploreSyncWorker::class.java)
+    }
+
+    @SuppressLint("CommitPrefEdits")
+    override suspend fun doWork(): Result = withContext(Dispatchers.IO) {
+        log.info("sync explore db started")
+
+        var localDataTimestamp = 0L
+        var remoteDataTimestamp = 0L
+        try {
+            syncStatus.setSyncProgress(0.0)
+
+            val timeInMillis = measureTimeMillis {
+                if (exploreRepository.localTimestamp == 0L) {
+                    // force data preloading for fresh installs
+                    AppExploreDatabase.getAppDatabase(
+                        appContext,
+                        config,
+                        exploreRepository
+                    )
+                }
+
+                localDataTimestamp = exploreRepository.localTimestamp
+                log.info("local data timestamp: $localDataTimestamp (${Date(localDataTimestamp)})")
+
+                remoteDataTimestamp = exploreRepository.getRemoteTimestamp()
+                log.info("remote data timestamp: $remoteDataTimestamp (${Date(remoteDataTimestamp)})")
+
+                if (localDataTimestamp >= remoteDataTimestamp) {
+                    log.info("explore db is up to date, nothing to sync")
+                    syncStatus.setSyncProgress(100.0)
+                    return@withContext Result.success()
+                }
+                syncStatus.setSyncProgress(10.0)
+
+                exploreRepository.download()
+
+                syncStatus.setSyncProgress(80.0)
+
+                AppExploreDatabase.forceUpdate(
+                    appContext,
+                    config,
+                    exploreRepository
+                )
+            }
+
+            log.info("sync explore db finished, took $timeInMillis ms")
+
+            syncStatus.setSyncProgress(100.0)
+
+        } catch (ex: FirebaseNetworkException) {
+            log.warn("sync explore no network", ex)
+            syncStatus.setSyncError(ex)
+            return@withContext Result.failure()
+        } catch (ex: Exception) {
+            analytics.logError(ex, "syncing from $localDataTimestamp, $remoteDataTimestamp")
+            log.error("sync explore db crashed ${ex.message}", ex)
+            syncStatus.setSyncError(ex)
+            return@withContext Result.failure()
+        }
+
+        return@withContext Result.success()
+    }
+}
