+++ wallet/src/de/schildbach/wallet/WalletApplication.java
@@ -76,6 +76,7 @@ import org.dash.wallet.common.InteractionAwareActivity;
 import org.dash.wallet.common.WalletDataProvider;
 import org.dash.wallet.common.transactions.TransactionFilter;
 import org.dash.wallet.common.transactions.TransactionWrapper;
+import org.dash.wallet.features.exploredash.ExploreSyncWorker;
 import org.dash.wallet.integration.liquid.data.LiquidClient;
 import org.dash.wallet.integration.liquid.data.LiquidConstants;
 import org.dash.wallet.integration.uphold.api.UpholdClient;
@@ -231,7 +232,6 @@ public class WalletApplication extends BaseWalletApplication
     }
 
     private void syncExploreData() {
-
         OneTimeWorkRequest syncDataWorkRequest =
                 new OneTimeWorkRequest.Builder(ExploreSyncWorker.class)
                         .setBackoffCriteria(
